pyproject.toml
[tool.poetry]
name = "permissions_system"
version = "0.2"
description = ""
authors = ["Shubham Biswas <shubhambiswas.zero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
databases = "^0.3.2"
sqlalchemy = "^1.3.16"
gino = "^1.0.0"
twine = "^3.1.1"
wheel = "^0.34.2"
fastapi = "^0.59.0"
python-dotenv = "^0.14.0"
uvicorn = "^0.11.5"
pyjwt = "^1.7.1"
bcrypt = "^3.1.7"
passlib = "^1.7.2"
gino-starlette = "^0.1.1"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
email_validator = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
